requirements.txt
python-dotenv
langchain
langchain_openai
openai
langchain-community
duckduckgo-search
routellm
streamlit

